Fix JWT token issuer claim validation
- Added issuer("agora.feedback") claim to all JWT tokens generated by JwtService
- This fixes "Token issued to client quarkus-app is not active" error when validating tokens
- Tokens now match the configured issuer in application.yaml (mp.jwt.verify.issuer)
- Applied to both token generation methods (with and without custom expiration)
- This enables proper JWT validation in authenticated endpoints

The fix ensures that when Quarkus validates JWT tokens via @Authenticated annotation,
the issuer claim in the token matches the configured issuer, preventing validation failures.

+++ src/main/resources/db/migration/V1__initial_schemas.sql
@@ -11,7 +11,7 @@ CREATE TABLE "user" (
     discord_username VARCHAR(255) NOT NULL,
     name VARCHAR(255) NOT NULL UNIQUE,
     avatar_url VARCHAR(255),
-    reputation_score SMALLINT NOT NULL DEFAULT 0,
+    reputation_score INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL,
     updated_at TIMESTAMP WITH TIME ZONE NOT NULL,
     archived_at TIMESTAMP WITH TIME ZONE,
